# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "balldontlie API",
        "description": "Teams, Players and Games management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get a list of players",
                "operationId": "SearchPlayer",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Create one player",
                "operationId": "CreatePlayer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "position": {
                                                "type": "string"
                                            },
                                            "height": {
                                                "type": "string"
                                            },
                                            "weight": {
                                                "type": "string"
                                            },
                                            "jersey_number": {
                                                "type": "string"
                                            },
                                            "college": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            },
                                            "draft_year": {
                                                "type": "integer"
                                            },
                                            "draft_round": {
                                                "type": "integer"
                                            },
                                            "draft_number": {
                                                "type": "integer"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "external_player_id": null,
                                        "first_name": "Jogador",
                                        "last_name": "numero 1",
                                        "position": "G",
                                        "height": "6-6",
                                        "weight": "190",
                                        "jersey_number": "8",
                                        "college": "Engenharia",
                                        "country": "Spain",
                                        "draft_year": "2013",
                                        "draft_round": 2,
                                        "draft_number": 32,
                                        "team_id": 1,
                                        "created_at": "2024-06-24T00:27:47.000000Z",
                                        "updated_at": "2024-06-24T00:29:02.000000Z"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/v1/players/{playerId}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get a one players",
                "operationId": "GetPlayerDetails",
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Update one player",
                "operationId": "UpdatePlayersDetails",
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "first_name": {
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "type": "string"
                                            },
                                            "position": {
                                                "type": "string"
                                            },
                                            "height": {
                                                "type": "string"
                                            },
                                            "weight": {
                                                "type": "string"
                                            },
                                            "jersey_number": {
                                                "type": "string"
                                            },
                                            "college": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            },
                                            "draft_year": {
                                                "type": "integer"
                                            },
                                            "draft_round": {
                                                "type": "integer"
                                            },
                                            "draft_number": {
                                                "type": "integer"
                                            },
                                            "team_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "external_player_id": null,
                                        "first_name": "Jogador",
                                        "last_name": "numero 1",
                                        "position": "G",
                                        "height": "6-6",
                                        "weight": "190",
                                        "jersey_number": "8",
                                        "college": "Engenharia",
                                        "country": "Spain",
                                        "draft_year": "2013",
                                        "draft_round": 2,
                                        "draft_number": 32,
                                        "team_id": 1,
                                        "created_at": "2024-06-24T00:27:47.000000Z",
                                        "updated_at": "2024-06-24T00:29:02.000000Z"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Delete a one player",
                "operationId": "DeletePlayerDetails",
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "path",
                        "description": "Player ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}